hyper roomba

+++ lib/engine/src/nodes/camera3d_node.cpp
@@ -6,7 +6,7 @@ namespace cathedral::engine
 {
     camera3d_node::camera3d_node(scene& scn, const std::string& name, scene_node* parent)
         : node(scn, name, parent)
-        , _camera(60, 16.0f / 9.0f, 0.01f, 100.0f)
+        , _camera(60, 16.0F / 9.0F, 0.01F, 100.0F)
     {
     }
 
@@ -18,7 +18,7 @@ namespace cathedral::engine
     void camera3d_node::tick([[maybe_unused]] double deltatime)
     {
         const auto surf_size = _scene.get_renderer().vkctx().get_surface_size();
-        const float aspect_ratio = static_cast<float>(surf_size.x) / surf_size.y;
+        const float aspect_ratio = static_cast<float>(surf_size.x) / static_cast<float>(surf_size.y);
         _camera.set_position(_local_transform.position());
         _camera.set_rotation(_local_transform.rotation());
         _camera.set_aspect_ratio(aspect_ratio);
